Wrap scan output in struct
A limitation of serverfull-gateway is that it can only handle objects
at the top level and is unable to handle top level arrays. This was done
intentionally to motivate us to use objects so that if we ever directly
pass through the function body it does not introduce a security issue.